+++ lib/core/llm_error.dart
@@ -1,3 +1,1 @@
-export '../src/compatibility/http/dio_error_handler.dart'
-    show DioErrorHandler, HttpErrorMapper;
 export '../src/compatibility/llm_error_types.dart';