+++ lib/providers/deepseek/error_handler.dart
@@ -1,6 +1,7 @@
 import 'package:llm_dart_transport/dio.dart';
 
 import '../../core/llm_error.dart';
+import '../../src/compatibility/http/dio_error_handler.dart';
 
 /// DeepSeek-specific error handler
 ///