+++ lib/providers/ollama/client.dart
@@ -10,6 +10,7 @@ import 'package:llm_dart_transport/llm_dart_transport.dart'
         bindDioCancellation;
 
 import '../../core/llm_error.dart';
+import '../../src/compatibility/http/dio_error_handler.dart';
 import '../../src/compatibility/http/http_response_handler.dart';
 import 'config.dart';
 import 'dio_strategy.dart';