+++ lib/src/compatibility/providers/anthropic/client.dart
@@ -3,8 +3,8 @@ import 'package:llm_dart_transport/llm_dart_transport.dart'
     show decodeDioResponseTextStream, Logger, ProviderDioClientFactory;
 
 import '../../../../core/cancellation.dart';
-import '../../../../core/llm_error.dart';
 import '../../../../providers/anthropic/config.dart';
+import '../../http/dio_error_handler.dart';
 import '../../http/dio_request_executor.dart';
 import '../../http/http_response_handler.dart';
 import 'dio_strategy.dart';